Add files via upload
After implementing this project I gave proper time to each and every module to test it, Black Box technique was used to test the system. Functions were tested by providing them with input and analysing the output.
I. Recognize the functions that the website module is anticipated to carry out. 
II. Make input data based on the module’s definition. 
III. Get the output based on the module’s specification. 
IV. Run the test cases. 
V. The last step is to collate the expected and actual output.